requirements.txt
streamlit
python-dotenv
google-generativeai
Pillow
matplotlib
numpy
SpeechRecognition
googletrans==4.0.0-rc1
